backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
openai==1.50.0
python-dotenv==1.0.1
numpy>=2.1.0
streamlit>=1.38.0
streamlit-folium>=0.22.0
folium>=0.17.0
plotly>=5.22.0
streamlit-js-eval>=0.1.7
extra-streamlit-components>=0.1.71
